refactor/#47: AuthImpl을 infrastructure로 이동
